fix: code clean up
