+++ LegalLead.PublicData.Search/FormMain.cs
@@ -24,7 +24,20 @@ namespace LegalLead.PublicData.Search
     [System.Diagnostics.CodeAnalysis.SuppressMessage("Reliability", "CA2007:Consider calling ConfigureAwait on the awaited task", Justification = "<Pending>")]
     public partial class FormMain : Form
     {
+        #region Private Members
+
+        private readonly string SubmitButtonText;
+
+        #endregion
+
+        #region Custom Properties
+
         protected WebFetchResult CaseData { get; set; }
+        private FindDbRequest CurrentRequest { get; set; } = null;
+
+        #endregion
+        
+        #region Constructor
 
         public FormMain()
         {
@@ -41,7 +54,11 @@ namespace LegalLead.PublicData.Search
             SetDentonStatusLabelFromSetting();
             SetStatus(StatusType.Ready);
         }
-        private readonly string SubmitButtonText;
+
+        #endregion
+
+        #region Form Event Handlers
+
         private void FormMain_Shown(object sender, EventArgs e)
         {
             SetInteraction(false);
@@ -64,17 +81,92 @@ namespace LegalLead.PublicData.Search
                     break;
             }
         }
-        private static string GetAccountIndexes()
+
+        private void FormMain_FormClosing(object sender, FormClosingEventArgs e)
         {
-            var container = SessionPersistenceContainer.GetContainer;
-            var instance = container.GetInstance<ISessionPersistance>(ApiHelper.ApiMode);
-            return instance.GetAccountPermissions();
+            const string processNames = "chromedriver,geckodriver";
+            KillProcess(processNames);
         }
 
-        private static bool IsAccountAdmin()
+        private void FormMain_Load(object sender, EventArgs e)
         {
-            return GetAccountIndexes().Equals("-1");
+            // method is intentionally left blank
         }
+
+        private void ExportDataToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            if (CaseData == null)
+            {
+                ShowNoDataErrorBox();
+                return;
+            }
+            if (string.IsNullOrEmpty(CaseData.Result))
+            {
+                ShowNoDataErrorBox();
+            }
+        }
+
+        private void ButtonDentonSetting_Click(object sender, EventArgs e)
+        {
+
+            var sourceId = ((WebNavigationParameter)cboWebsite.SelectedItem).Id;
+            switch (sourceId)
+            {
+                case (int)SourceType.DentonCounty:
+                    using (var credential = new FormDentonSetting())
+                    {
+                        credential.Icon = Icon;
+                        credential.ShowDialog(this);
+                        SetDentonStatusLabelFromSetting();
+                    }
+                    break;
+                default:
+                    using (var result = new FormCredential())
+                    {
+                        result.Icon = Icon;
+                        result.ShowDialog(this);
+                    }
+                    break;
+            }
+        }
+
+        private void ToolStripSplitButton1_ButtonClick(object sender, EventArgs e)
+        {
+            var settings = new FormSettings() { StartPosition = FormStartPosition.CenterParent };
+            settings.ShowDialog();
+        }
+
+        #endregion
+
+        #region Private Helper Methods
+
+        private void SetUserName()
+        {
+            var username = string.Empty;
+            try
+            {
+                username = GetUserName();
+                if (string.IsNullOrEmpty(username)) return;
+            }
+            finally
+            {
+                var isVisible = !string.IsNullOrEmpty(username);
+                var txt = isVisible ? $" | {username}" : string.Empty;
+                tsUserName.Text = txt;
+                tsUserName.Visible = isVisible;
+            }
+        }
+
+        private void SetInteraction(bool isEnabled)
+        {
+            if (!isEnabled) TryHideProgress();
+            cboWebsite.Enabled = isEnabled;
+            button1.Enabled = isEnabled;
+            tsSettingMenuButton.Visible = isEnabled;
+            dteEnding.Enabled = isEnabled;
+            dteStart.Enabled = isEnabled;
+        }
+
         private void FilterWebSiteByAccount()
         {
             try
@@ -118,48 +210,75 @@ namespace LegalLead.PublicData.Search
             }
         }
 
-        private static string GetUserName()
+
+        private void TryHideProgress()
         {
-            var container = AuthenicationContainer.GetContainer;
-            var userservice = container.GetInstance<SessionUserPersistence>();
-            return userservice.GetUserName();
-        }
-        private void SetUserName()
-        {
-            var username = string.Empty;
             try
             {
-                username = GetUserName();
-                if (string.IsNullOrEmpty(username)) return;
+                HideProgress();
             }
-            finally
+            catch
             {
-                var isVisible = !string.IsNullOrEmpty(username);
-                var txt = isVisible ? $" | {username}" : string.Empty;
-                tsUserName.Text = txt;
-                tsUserName.Visible = isVisible;
+                Invoke(HideProgress);
             }
         }
-        private void SetInteraction(bool isEnabled)
+
+        private void ClearProgressDate()
         {
-            if (!isEnabled) TryHideProgress();
-            cboWebsite.Enabled = isEnabled;
-            button1.Enabled = isEnabled;
-            tsSettingMenuButton.Visible = isEnabled;
-            dteEnding.Enabled = isEnabled;
-            dteStart.Enabled = isEnabled;
+            try
+            {
+                lbProgressDate.Text = string.Empty;
+            }
+            catch
+            {
+                Invoke(() => { lbProgressDate.Text = string.Empty; });
+            }
         }
 
-        private void FormMain_FormClosing(object sender, FormClosingEventArgs e)
+        private void TryShowProgress(int min, int max, int current, string dateIndication)
         {
-            const string processNames = "chromedriver,geckodriver";
-            KillProcess(processNames);
+            try
+            {
+                ShowProgress(min, max, current, dateIndication);
+            }
+            catch
+            {
+                Invoke(() => { ShowProgress(min, max, current, dateIndication); });
+            }
         }
 
-        protected static T GetObject<T>(object item)
+        private void HideProgress()
         {
-            return (T)item;
+            if (!string.IsNullOrEmpty(lbProgressDate.Text)) return;
+            progressBar1.Visible = false;
+            labelProgress.Visible = false;
+            tableLayoutPanel1.RowStyles[8].Height = 0;
         }
+
+        private void ShowProgress(int min, int max, int current, string dateIndication = "")
+        {
+            labelProgress.Visible = true;
+            lbProgressDate.Visible =
+                !string.IsNullOrEmpty(lbProgressDate.Text) ||
+                !string.IsNullOrEmpty(dateIndication);
+            if (dateIndication == "hide")
+            {
+                lbProgressDate.Text = string.Empty;
+                lbProgressDate.Visible = false;
+                TryHideProgress();
+                return;
+            }
+            if (!string.IsNullOrEmpty(dateIndication)) lbProgressDate.Text = dateIndication;
+            ControlExtensions.Suspend(progressBar1);
+            tableLayoutPanel1.RowStyles[8].Height = 40;
+            progressBar1.Visible = false;
+            progressBar1.Minimum = min;
+            progressBar1.Maximum = max;
+            progressBar1.Value = current;
+            progressBar1.Visible = true;
+            ControlExtensions.Resume();
+        }
+
         private void SetStatus(StatusType status)
         {
             var v = StatusHelper.GetStatus(status);
@@ -191,207 +310,17 @@ namespace LegalLead.PublicData.Search
             }
         }
 
-        private void Button1_Click(object sender, EventArgs e)
+        private void SetStatusFromOffThread(StatusState v, StatusType status = StatusType.None)
         {
-            string driverNames = string.Concat("geckodriver,", CommonKeyIndexes.ChromeDriver);
-            try
+            this.Invoke(new Action(() =>
             {
-                KillProcess(driverNames);
-                SetStatus(StatusType.Running);
-                using var hider = new HideProcessWindowHelper();
-                if (!ValidateCustom())
-                {
-                    SetStatus(StatusType.Ready);
-                    return;
-                }
-                txConsole.Text = string.Empty;
-                ProcessStartingMessage();
-                var startDate = dteStart.Value.Date;
-                var endingDate = dteEnding.Value.Date;
-                var siteData = (WebNavigationParameter)(cboWebsite.SelectedItem);
-                var sourceMember = (SourceType)siteData.Id;
-                CurrentRequest = sourceMember.GetDbRequest(this, startDate);
-                var searchItem = new SearchResult
-                {
-                    Id = GetObject<List<SearchResult>>(Tag).Count + 1,
-                    Website = siteData.Name,
-                    EndDate = endingDate.ToShortDateString(),
-                    StartDate = startDate.ToShortDateString(),
-                    SearchDate = DateTime.Now.ToShortDateString() + " - " + DateTime.Now.ToShortTimeString(),
-                };
-                searchItem.Search = $"{searchItem.SearchDate} : {searchItem.Website} from {searchItem.StartDate} to {searchItem.EndDate}";
-                switch (siteData.Id)
-                {
-                    case (int)SourceType.BexarCounty:
-                        BexarButtonExecution(siteData, searchItem);
-                        break;
-                    case (int)SourceType.DallasCounty:
-                        DallasButtonExecution(siteData, searchItem);
-                        break;
-                    case (int)SourceType.TravisCounty:
-                        TravisButtonExecution(siteData, searchItem, startDate, endingDate);
-                        break;
-                    case (int)SourceType.HarrisCivil:
-                        var cbindx = GetCaseSelectionIndex(cboCaseType.SelectedItem);
-                        if (cbindx == 0) NonDallasButtonExecution(siteData, searchItem, startDate, endingDate);
-                        else CommonButtonExecution(siteData, searchItem);
-                        break;
-                    case (int)SourceType.HidalgoCounty:
-                    case (int)SourceType.ElPasoCounty:
-                    case (int)SourceType.FortBendCounty:
-                    case (int)SourceType.WilliamsonCounty:
-                    case (int)SourceType.GraysonCounty:
-                        CommonButtonExecution(siteData, searchItem);
-                        break;
-                    default:
-                        NonDallasButtonExecution(siteData, searchItem, startDate, endingDate);
-                        break;
-                }
-
-            }
-            catch (Exception ex)
-            {
-                SetStatus(StatusType.Error);
-                Console.WriteLine(CommonKeyIndexes.UnexpectedErrorOccurred);
-                Console.WriteLine(ex.Message);
-
-                Console.WriteLine(CommonKeyIndexes.DashedLine);
-            }
-            finally
-            {
-                KillProcess(driverNames);
-                CurrentRequest = null;
-            }
+                toolStripStatus.Text = string.Format(CultureInfo.CurrentCulture, "{0}", v.Name);
+                toolStripStatus.ForeColor = v.Color;
+                var enabled = status != StatusType.Running;
+                SetControlEnabledState(enabled);
+                Refresh();
+            }));
         }
-
-        private static int GetCaseSelectionIndex(object selectedItem)
-        {
-            var fallback = 0;
-            if (selectedItem == null) return fallback;
-            if (selectedItem is not Option ddp) return fallback;
-            if (!ddp.Name.Contains("criminal")) return fallback;
-            return 1;
-        }
-
-        private void CommonButtonExecution(WebNavigationParameter siteData, SearchResult searchItem)
-        {
-            var index = cboSearchType.SelectedIndex;
-            var searchType = DallasSearchProcess.GetCourtName(index);
-            var keys = new List<WebNavigationKey> {
-                new() { Name = "StartDate", Value = searchItem.StartDate},
-                new() { Name = "EndDate", Value = searchItem.EndDate },
-                new() { Name = "CourtType", Value = searchType }
-            };
-            var wb = new WebNavigationParameter { Keys = keys };
-            IWebInteractive dweb = siteData.Id switch
-            {
-                30 => new HccUiInteractive(wb),
-                130 => new GraysonUiInteractive(wb),
-                120 => new WilliamsonUiInteractive(wb),
-                110 => new FortBendUiInteractive(wb),
-                100 => new ElPasoUiInteractive(wb),
-                _ => new HidalgoUiInteractive(wb)
-            };
-            var indx = siteData.Id == 30 ? 1 : 0;
-            _ = Task.Run(async () =>
-            {
-                await ProcessAsync(dweb, siteData, searchItem, indx);
-            }).ConfigureAwait(true);
-        }
-
-        private void BexarButtonExecution(WebNavigationParameter siteData, SearchResult searchItem)
-        {
-            var index = cboSearchType.SelectedIndex;
-            var searchType = DallasSearchProcess.GetCourtName(index);
-            var keys = new List<WebNavigationKey> {
-                new() { Name = "StartDate", Value = searchItem.StartDate},
-                new() { Name = "EndDate", Value = searchItem.EndDate },
-                new() { Name = "CourtType", Value = searchType }
-            };
-            var wb = new WebNavigationParameter { Keys = keys };
-            var dweb = new BexarUiInteractive(wb);
-            _ = Task.Run(async () =>
-            {
-                await ProcessAsync(dweb, siteData, searchItem);
-            }).ConfigureAwait(true);
-        }
-
-        private void DallasButtonExecution(WebNavigationParameter siteData, SearchResult searchItem)
-        {
-            var container = ActionSettingContainer.GetContainer;
-            var index = cboSearchType.SelectedIndex;
-            var searchType = DallasSearchProcess.GetCourtName(index);
-            var keys = new List<WebNavigationKey> {
-                new() { Name = "StartDate", Value = searchItem.StartDate},
-                new() { Name = "EndDate", Value = searchItem.EndDate },
-                new() { Name = "CourtType", Value = searchType }
-            };
-            var wb = new WebNavigationParameter { Keys = keys };
-            var dweb = new DallasUiInteractive(wb);
-            var dbweb = new UiDbInteractive(
-                dweb,
-                container.GetInstance<IRemoteDbHelper>(),
-                CurrentRequest);
-            _ = Task.Run(async () =>
-            {
-                await ProcessAsync(dbweb, siteData, searchItem);
-            }).ConfigureAwait(true);
-        }
-
-        private void TravisButtonExecution(WebNavigationParameter siteData, SearchResult searchItem, DateTime startDate, DateTime endingDate)
-        {
-            var dto = (DropDown)cboSearchType.SelectedItem;
-            var txt = dto.Name.Split(' ')[0];
-            var searchType = txt.ToUpper(CultureInfo.CurrentCulture).Trim();
-            var search = new TravisSearchProcess();
-            search.Search(startDate, endingDate, searchType);
-            var dweb = search.GetUiInteractive();
-            _ = Task.Run(async () =>
-            {
-                await ProcessAsync(dweb, siteData, searchItem);
-            }).ConfigureAwait(true);
-        }
-
-        private void NonDallasButtonExecution(WebNavigationParameter siteData, SearchResult searchItem, DateTime startDate, DateTime endingDate)
-        {
-            const StringComparison ccic = StringComparison.CurrentCultureIgnoreCase;
-            var isDentonCounty = siteData.Id == (int)SourceType.DentonCounty;
-            var keys = siteData.Keys;
-            var isDistrictSearch = keys.Find(x =>
-                x.Name.Equals(CommonKeyIndexes.DistrictSearchType, // "DistrictSearchType"
-                ccic)) != null;
-            var criminalToggle = keys.Find(x =>
-                x.Name.Equals(CommonKeyIndexes.CriminalCaseInclusion,
-                ccic));
-            if (isDentonCounty && criminalToggle != null)
-            {
-                criminalToggle.Value = isDistrictSearch ?
-                    CommonKeyIndexes.NumberZero :
-                    CommonKeyIndexes.NumberOne;
-            }
-
-            if (!isDentonCounty && criminalToggle != null)
-            {
-                criminalToggle.Value = CommonKeyIndexes.NumberOne;
-            }
-            var collinIndex = (int)SourceType.CollinCounty;
-            if (siteData.Id == collinIndex)
-            {
-                siteData.Keys.Add(new() { Name = "StartDate", Value = $"{startDate:d}" });
-                siteData.Keys.Add(new() { Name = "EndDate", Value = $"{endingDate:d}" });
-                siteData.Keys.Add(new() { Name = "CourtType", Value = "Civil" });
-            }
-            IWebInteractive webmgr =
-            siteData.Id == collinIndex ?
-            new CollinUiInterative(siteData, startDate, endingDate) :
-            WebFetchingProvider.GetInteractive(siteData, startDate, endingDate);
-            var cbindx = siteData.Id != 30 ? 0 : GetCaseSelectionIndex(cboCaseType.SelectedItem);
-            _ = Task.Run(async () =>
-            {
-                await ProcessAsync(webmgr, siteData, searchItem, cbindx);
-            }).ConfigureAwait(true);
-        }
-
 
         private void SortWebsites()
         {
@@ -408,17 +337,23 @@ namespace LegalLead.PublicData.Search
             cboWebsite.SelectedIndex = itemIndex;
         }
 
-        private void ExportDataToolStripMenuItem_Click(object sender, EventArgs e)
+
+        #endregion
+
+        #region Static Methods
+
+        protected static T GetObject<T>(object item)
         {
-            if (CaseData == null)
-            {
-                ShowNoDataErrorBox();
-                return;
-            }
-            if (string.IsNullOrEmpty(CaseData.Result))
-            {
-                ShowNoDataErrorBox();
-            }
+            return (T)item;
+        }
+
+        private static int GetCaseSelectionIndex(object selectedItem)
+        {
+            var fallback = 0;
+            if (selectedItem == null) return fallback;
+            if (selectedItem is not Option ddp) return fallback;
+            if (!ddp.Name.Contains("criminal")) return fallback;
+            return 1;
         }
 
         private static void ShowNoDataErrorBox()
@@ -429,33 +364,22 @@ namespace LegalLead.PublicData.Search
                 MessageBoxIcon.Error);
         }
 
-        private void FormMain_Load(object sender, EventArgs e)
+        private static string GetUserName()
         {
-            // method is intentionally left blank
+            var container = AuthenicationContainer.GetContainer;
+            var userservice = container.GetInstance<SessionUserPersistence>();
+            return userservice.GetUserName();
+        }
+        private static bool IsAccountAdmin()
+        {
+            return GetAccountIndexes().Equals("-1");
         }
 
-        private void ButtonDentonSetting_Click(object sender, EventArgs e)
+        private static string GetAccountIndexes()
         {
-
-            var sourceId = ((WebNavigationParameter)cboWebsite.SelectedItem).Id;
-            switch (sourceId)
-            {
-                case (int)SourceType.DentonCounty:
-                    using (var credential = new FormDentonSetting())
-                    {
-                        credential.Icon = Icon;
-                        credential.ShowDialog(this);
-                        SetDentonStatusLabelFromSetting();
-                    }
-                    break;
-                default:
-                    using (var result = new FormCredential())
-                    {
-                        result.Icon = Icon;
-                        result.ShowDialog(this);
-                    }
-                    break;
-            }
+            var container = SessionPersistenceContainer.GetContainer;
+            var instance = container.GetInstance<ISessionPersistance>(ApiHelper.ApiMode);
+            return instance.GetAccountPermissions();
         }
 
         private static void KillProcess(string processName)
@@ -476,7 +400,6 @@ namespace LegalLead.PublicData.Search
             }
         }
 
-
         private static bool IsEmpty(WebFetchResult caseData)
         {
             if (caseData == null) return true;
@@ -484,123 +407,11 @@ namespace LegalLead.PublicData.Search
             return caseData.CaseList.Equals(emptyCases, StringComparison.OrdinalIgnoreCase);
         }
 
-        private async Task ProcessAsync(
-            IWebInteractive webmgr,
-            WebNavigationParameter siteData,
-            SearchResult searchItem,
-            int caseSelectionIndex = 0)
-        {
-            try
-            {
-                if (IsAccountAdmin())
-                {
-                    var displayMode = SettingsWriter.GetSettingOrDefault("browser", "Open Headless:", true);
-                    if (!displayMode) { webmgr.DriverReadHeadless = false; }
-                }
-                webmgr.ReportProgress = TryShowProgress;
-                webmgr.ReportProgessComplete = TryHideProgress;
-                var interactive = GetDbInteractive(webmgr, siteData.Id);
-                CaseData = await Task.Run(() =>
-                {
-                    return interactive.Fetch();
-                }).ConfigureAwait(true);
 
-                var nonactors = new List<int> {
-                    (int)SourceType.CollinCounty,
-                    (int)SourceType.DallasCounty,
-                    (int)SourceType.TravisCounty,
-                    (int)SourceType.TarrantCounty,
-                    (int)SourceType.BexarCounty,
-                    (int)SourceType.HarrisCivil,
-                    (int)SourceType.HidalgoCounty,
-                    (int)SourceType.ElPasoCounty,
-                    (int)SourceType.FortBendCounty,
-                    (int)SourceType.WilliamsonCounty,
-                    (int)SourceType.GraysonCounty,
-                };
+        #endregion
 
-                ProcessEndingMessage();
-                SetStatus(StatusType.Finished);
-                KillProcess(CommonKeyIndexes.ChromeDriver);
-                if (CaseData == null)
-                {
-                    throw new KeyNotFoundException(CommonKeyIndexes.NoDataFoundFromCaseExtract);
-                }
-                if (string.IsNullOrEmpty(CaseData.Result))
-                {
-                    throw new KeyNotFoundException(CommonKeyIndexes.NoDataFoundFromCaseExtract);
-                }
-                if (IsEmpty(CaseData))
-                {
-                    throw new KeyNotFoundException(CommonKeyIndexes.NoDataFoundFromCaseExtract);
-                }
-                CaseData.WebsiteId = siteData.Id;
-                // write search count to api
-                var count = CaseData.PeopleList.Count;
-                if (count > 0)
-                {
-                    var member = (SourceType)siteData.Id;
-                    var userName = GetUserName();
-                    var searchRange = $"{webmgr.StartDate:d} to {webmgr.EndingDate:d}";
-                    if (string.IsNullOrWhiteSpace(userName)) { userName = "unknown"; }
-                    var countyName = member.GetCountyName();
-                    UsageIncrementer.IncrementUsage(userName, countyName, count, searchRange);
-                    UsageReader.WriteUserRecord();
-                }
+        #region Private Static Fields
 
-                var isHarrisCriminal = caseSelectionIndex == 1 && siteData.Id == (int)SourceType.HarrisCivil;
-                if (!isHarrisCriminal && !nonactors.Contains(siteData.Id))
-                {
-                    ExcelWriter.WriteToExcel(CaseData);
-                }
-                searchItem.ResultFileName = CaseData.Result;
-                searchItem.IsCompleted = true;
-                searchItem.MoveToCommon();
-                GetObject<List<SearchResult>>(Tag).Add(searchItem);
-                ComboBox_DataSourceChanged(null, null);
-
-                var result = MessageBox.Show(
-                    CommonKeyIndexes.CaseExtractCompleteWouldYouLikeToView,
-                    CommonKeyIndexes.DataExtractSuccess,
-                    MessageBoxButtons.YesNo,
-                    MessageBoxIcon.Question);
-                if (result != DialogResult.Yes)
-                {
-                    SetStatus(StatusType.Ready);
-                    return;
-                }
-                TryOpenExcel();
-                SetStatus(StatusType.Ready);
-            }
-            catch (Exception ex)
-            {
-                SetStatus(StatusType.Error);
-                Console.WriteLine(CommonKeyIndexes.UnexpectedErrorOccurred);
-                Console.WriteLine(ex.Message);
-
-                Console.WriteLine(CommonKeyIndexes.DashedLine);
-            }
-            finally
-            {
-                KillProcess(CommonKeyIndexes.ChromeDriver);
-                KillProcess("geckodriver");
-                ClearProgressDate();
-                TryHideProgress();
-            }
-        }
-
-        private void SetStatusFromOffThread(StatusState v, StatusType status = StatusType.None)
-        {
-            this.Invoke(new Action(() =>
-            {
-                toolStripStatus.Text = string.Format(CultureInfo.CurrentCulture, "{0}", v.Name);
-                toolStripStatus.ForeColor = v.Color;
-                var enabled = status != StatusType.Running;
-                SetControlEnabledState(enabled);
-                Refresh();
-            }));
-        }
-        private FindDbRequest CurrentRequest { get; set; } = null;
         private static readonly SessionUsagePersistence UsageIncrementer
             = SessionPersistenceContainer.GetContainer
             .GetInstance<SessionUsagePersistence>();
@@ -609,101 +420,9 @@ namespace LegalLead.PublicData.Search
             .GetInstance<SessionMonthToDatePersistence>();
 
 
-        private void ToolStripSplitButton1_ButtonClick(object sender, EventArgs e)
-        {
-            var settings = new FormSettings() { StartPosition = FormStartPosition.CenterParent };
-            settings.ShowDialog();
-        }
-
-        private void TryHideProgress()
-        {
-            try
-            {
-                HideProgress();
-            }
-            catch
-            {
-                Invoke(HideProgress);
-            }
-        }
-
-        private void ClearProgressDate()
-        {
-            try
-            {
-                lbProgressDate.Text = string.Empty;
-            }
-            catch
-            {
-                Invoke(() => { lbProgressDate.Text = string.Empty; });
-            }
-        }
-        private void TryShowProgress(int min, int max, int current, string dateIndication)
-        {
-            try
-            {
-                ShowProgress(min, max, current, dateIndication);
-            }
-            catch
-            {
-                Invoke(() => { ShowProgress(min, max, current, dateIndication); });
-            }
-        }
-
-        private void HideProgress()
-        {
-            if (!string.IsNullOrEmpty(lbProgressDate.Text)) return;
-            progressBar1.Visible = false;
-            labelProgress.Visible = false;
-            tableLayoutPanel1.RowStyles[8].Height = 0;
-        }
-        private void ShowProgress(int min, int max, int current, string dateIndication = "")
-        {
-            labelProgress.Visible = true;
-            lbProgressDate.Visible =
-                !string.IsNullOrEmpty(lbProgressDate.Text) ||
-                !string.IsNullOrEmpty(dateIndication);
-            if (dateIndication == "hide")
-            {
-                lbProgressDate.Text = string.Empty;
-                lbProgressDate.Visible = false;
-                TryHideProgress();
-                return;
-            }
-            if (!string.IsNullOrEmpty(dateIndication)) lbProgressDate.Text = dateIndication;
-            ControlExtensions.Suspend(progressBar1);
-            tableLayoutPanel1.RowStyles[8].Height = 40;
-            progressBar1.Visible = false;
-            progressBar1.Minimum = min;
-            progressBar1.Maximum = max;
-            progressBar1.Value = current;
-            progressBar1.Visible = true;
-            ControlExtensions.Resume();
-        }
-
-        private IWebInteractive GetDbInteractive(IWebInteractive interactive, int siteId)
-        {
-            if (interactive is UiDbInteractive db) return db;
-            if (interactive is HccUiInteractive hcc) return hcc;
-            var container = ActionSettingContainer.GetContainer;
-            if (CurrentRequest == null)
-            {
-                var member = (SourceType)siteId;
-                CurrentRequest = member.GetDbRequest(this, DateTime.Now.Date);
-            }
-            var response = new UiDbInteractive(
-                interactive,
-                container.GetInstance<IRemoteDbHelper>(),
-                CurrentRequest)
-            {
-                ReportProgress = TryShowProgress,
-                ReportProgessComplete = TryHideProgress,
-                StartDate = dteStart.Value.Date,
-                EndingDate = dteEnding.Value.Date
-            };
-            response.Parameters = interactive.Parameters ?? new();
-            return response;
-        }
+        #endregion
+        
+        #region Classes
 
         private static class ControlExtensions
         {
@@ -721,5 +440,7 @@ namespace LegalLead.PublicData.Search
             }
 
         }
+
+        #endregion
     }
 }
