--- LegalLead.PublicData.Search/SetTextCallback.cs
@@ -1,6 +1,0 @@
-﻿namespace LegalLead.PublicData.Search
-{
-    internal class SetTextCallback
-    {
-    }
-}
