+++ Thompson.RecordSearch.Utility/Resources/credential_services_map.txt
@@ -1,5 +1,5 @@
 ﻿{
-	"web": "http://legalleadpermissionsapi-dev.us-east-2.elasticbeanstalk.com",
+	"web": "http://api.legallead.co",
 	"landings":
 	{
 		"login": "/app/login",
