+++ Thompson.RecordSearch.Utility/_html/credential_services_map.txt
@@ -1,5 +1,5 @@
 ﻿{
-	"web": "http://legalleadpermissionsapi-dev.us-east-2.elasticbeanstalk.com/app/get-county-code",
+	"web": "http://api.legallead.co/app/get-county-code",
 	"counties": 
 		[
 			{
